refactor: make mount_session  re-usable * move mount_session to honestnft_utils.misc * document honestnft_utils.misc
